+++ app/src/main/java/com/example/synema/view/screens/LoginScreen.kt
@@ -1,0 +1,80 @@
+package com.example.synema.view.screens
+
+import GradientBox
+import MoviePosterFrame
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawWithCache
+import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavHostController
+import coil.compose.AsyncImage
+import com.example.synema.R
+import com.example.synema.view.components.SynemaLogo
+
+
+@Composable
+     fun LoginScreen(navController : NavHostController) {
+        GradientBox(){
+            ContentContainer();
+        }
+    }
+
+    @Composable
+    private fun ContentContainer(){
+        Column (
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(14.dp)
+            ){
+            SynHeader()
+            MovieDisplay();
+        }
+    }
+
+    @Composable
+    private fun MovieDisplay(){
+        Row(
+            horizontalArrangement = Arrangement.Center,
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(250.dp)
+        ){
+            MoviePosterFrame(Arrangement.Bottom, "https://static.posters.cz/image/750/plakater/interstellar-ice-walk-i23290.jpg")
+            MoviePosterFrame(Arrangement.Center, "https://i.etsystatic.com/10683147/r/il/d4a024/4900691314/il_1080xN.4900691314_fu21.jpg")
+            MoviePosterFrame(Arrangement.Top, "https://www.hollywoodreporter.com/wp-content/uploads/2023/06/French-Film-Poster-Barbie-Warner-Bros..jpg?w=999")
+        }
+    }
+
+    @Composable
+    private fun SynHeader() {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 77.dp)
+
+        ) {
+            SynemaLogo()
+        }
+    }
